feat: enhance UI, performance and metrology
This commit introduces several key updates:

- Implemented CancellationToken via tokio::select! to abort heavy CPU tasks on tab switches without leaking resources.

- Added dynamic period selection (3M, 6M, 1Y, 2Y, 5Y, 10Y) to HistoricoTab.

- Improved UI legibility by removing low-contrast gray text in favor of crisp black/white modes.

- Added automated CSV exports for asset analytics (results/ directory).

- Resolved all linter warnings, dead code, and formatted via cargo fmt.

+++ src/provider/indices/ibovespa/options.rs
@@ -56,13 +56,13 @@ impl Options {
                 let date_str = dt.naive_utc().format("%d/%m/%Y").to_string();
 
                 let ibov = Ibov {
-                    timestamp: q.timestamp,
+                    timestamp: q.timestamp as u64,
                     adjclose: q.adjclose,
                     date: date_str,
                     open: q.open,
                     high: q.high,
                     low: q.low,
-                    volume: q.volume,
+                    volume: q.volume as u64,
                     close: q.close,
                 };
                 ibovs.push(ibov);
